+++ packages/core/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.8.2](https://github.com/Inist-CNRS/ezs/compare/@ezs/core@1.8.1...@ezs/core@1.8.2) (2020-04-17)
+
+**Note:** Version bump only for package @ezs/core
+
+
+
+
+
 ## [1.8.1](https://github.com/Inist-CNRS/ezs/compare/@ezs/core@1.8.0...@ezs/core@1.8.1) (2020-04-06)